backend/requirements.txt
Django==5.1.2
djangorestframework==3.16.1
django-cors-headers==4.8.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
djangorestframework-simplejwt==5.3.1
reportlab==4.4.4
Pillow==12.0.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
psutil==6.1.0

